+++ src/calculator/__tests__/Value.ts
@@ -11,6 +11,22 @@ it('should match tags', () => {
       }),
     ),
   ).toBe(true);
+  expect(
+    local_armour.augmentableBy(
+      new Stat({
+        id: 'local_evasion_and_energy_shield_+%',
+      }),
+    ),
+  ).toBe(false);
+
+  const local_shield = new Value([0, 0], ['local', 'block']);
+  expect(
+    local_shield.augmentableBy(
+      new Stat({
+        id: 'additional_block_%',
+      }),
+    ),
+  ).toBe(true);
 });
 
 it('should know about the difference between flat, inc and more', () => {