+++ CHANGELOG.md
@@ -22,6 +22,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
   It is not supposed to provide any functionallity upon application. 
   ([#42](https://github.com/eps1lon/poe-mods/pull/42))
 - Wrong mods for Abyss Jewels being listed. ([#45](https://github.com/eps1lon/poe-mods/pull/45))
+- Fixed item quality not affecting Armour, Evasion, Energy Shield and Physical 
+  Damage. Reported in [#38](https://github.com/eps1lon/poe-mods/issues/38) and
+  fixed in [#52](https://github.com/eps1lon/poe-mods/pull/52)
 
 ## [1.8.0](https://github.com/eps1lon/poe-mods/compare/v1.7.0...v1.8.0) (2018-03-04)